More update about the difficulity setting and fuel
So, I added the case to switch, so user can put case 1(easy), case 2(Normal), case 3(Hard). case 4(Custom)
so the easy is 5 pieces, normal is 10 pieces  and hard is 20 pieces, the Custom mode allows player to enter the pieces they want to collect, the amount of pieces might change if I change my mind. And about the fuel, it's just a if function, if user move successfully, the amount of fuel -1, until the fuel = 0, player lose